Crate narwhal_network

Modules

admin

anemo_ext

connectivity

epoch_filter

failpoints

metrics

Structs

CancelOnDropHandler

This adapter will make a [tokio::task::JoinHandle] abort its handled task when the handle is dropped.

RetryConfig

Retry configurations for establishing connections and sending messages. Determines the retry behaviour of requests, by setting the back off strategy used.

Traits

PrimaryToPrimaryRpc

PrimaryToWorkerRpc

ReliableNetwork

UnreliableNetwork

WorkerRpc

Functions

multiaddr_to_address

Attempts to convert a multiaddr of the form /[ip4,ip6,dns]/{}/udp/{port} into an anemo address