Trait narwhal_network::UnreliableNetwork

pub trait UnreliableNetwork<Request: Clone + Send + Sync> {
    type Response: Clone + Send + Sync;

    fn unreliable_send(
        &mut self,
        peer: NetworkPublicKey,
        message: &Request
    ) -> Result<JoinHandle<Result<Response<Self::Response>>>>;

    fn unreliable_broadcast(
        &mut self,
        peers: Vec<NetworkPublicKey>,
        message: &Request
    ) -> Vec<Result<JoinHandle<Result<Response<Self::Response>>>>> { ... }
}

Required Associated Types

type Response: Clone + Send + Sync

Required Methods

fn unreliable_send(
&mut self,
peer: NetworkPublicKey,
message: &Request
) -> Result<JoinHandle<Result<Response<Self::Response>>>>

Provided Methods

fn unreliable_broadcast(
&mut self,
peers: Vec<NetworkPublicKey>,
message: &Request
) -> Vec<Result<JoinHandle<Result<Response<Self::Response>>>>>

Broadcasts a message to all peers passed as an argument. The attempts to send individual messages are best effort and will not be retried.