Crate narwhal_network

Modules

admin

client

connectivity

epoch_filter

failpoints

metrics

Structs

CancelOnDropHandler

This adapter will make a [tokio::task::JoinHandle] abort its handled task when the handle is dropped.

RetryConfig

Retry configurations for establishing connections and sending messages. Determines the retry behaviour of requests, by setting the back off strategy used.

Traits

PrimaryToPrimaryRpc

PrimaryToWorkerClient

ReliableNetwork

WorkerRpc

WorkerToPrimaryClient